feat: add SWGOH string parsing utility and improve error handling
